+++ Changes.md
@@ -6,6 +6,7 @@
 API
 ---
 
+- PathColumn : Added `buttonPressSignal()`, `buttonReleaseSignal()` and `buttonDoubleClickSignal()`. These allow a PathColumn to implement its own event handling.
 - Capsule : Removed attempts to detect invalidated Capsules.
 
 Breaking Changes